+++ spec/dummy/config/application.rb
@@ -23,4 +23,3 @@ module Dummy
     # -- all .rb files in that directory are automatically loaded.
   end
 end
-