+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In Development
 
 1. Configure travis CI to use cached versions of gems.  (Issue #8)
+1. Correct rubocop issues.  (Issue #11)
 
 ## 1.0.1 (May 11, 2016)